[FSE - Templates]: Add info to templates created from wp suggestion
