+++ lib/compat/wordpress-5.9/block-template-utils.php
@@ -529,272 +529,6 @@ if ( ! function_exists( '_build_block_template_result_from_file' ) ) {
 	}
 }
 
-if ( ! function_exists( '_build_block_template_result_from_post' ) ) {
-	/**
-	 * Build a unified template object based a post Object.
-	 *
-	 * @param WP_Post $post Template post.
-	 *
-	 * @return Gutenberg_Block_Template|WP_Error Template.
-	 */
-	function _build_block_template_result_from_post( $post ) {
-		$default_template_types = get_default_block_template_types();
-		$terms                  = get_the_terms( $post, 'wp_theme' );
-
-		if ( is_wp_error( $terms ) ) {
-			return $terms;
-		}
-
-		if ( ! $terms ) {
-			return new WP_Error( 'template_missing_theme', __( 'No theme is defined for this template.', 'gutenberg' ) );
-		}
-
-		$origin = get_post_meta( $post->ID, 'origin', true );
-
-		$theme          = $terms[0]->name;
-		$has_theme_file = wp_get_theme()->get_stylesheet() === $theme &&
-			null !== _get_block_template_file( $post->post_type, $post->post_name );
-
-		$template                 = new Gutenberg_Block_Template();
-		$template->wp_id          = $post->ID;
-		$template->id             = $theme . '//' . $post->post_name;
-		$template->theme          = $theme;
-		$template->content        = $post->post_content;
-		$template->slug           = $post->post_name;
-		$template->source         = 'custom';
-		$template->origin         = ! empty( $origin ) ? $origin : null;
-		$template->type           = $post->post_type;
-		$template->description    = $post->post_excerpt;
-		$template->title          = $post->post_title;
-		$template->status         = $post->post_status;
-		$template->has_theme_file = $has_theme_file;
-		$template->is_custom      = true;
-		$template->author         = $post->post_author;
-
-		if ( 'wp_template' === $post->post_type && isset( $default_template_types[ $template->slug ] ) ) {
-			$template->is_custom = false;
-		}
-
-		if ( 'wp_template_part' === $post->post_type ) {
-			$type_terms = get_the_terms( $post, 'wp_template_part_area' );
-			if ( ! is_wp_error( $type_terms ) && false !== $type_terms ) {
-				$template->area = $type_terms[0]->name;
-			}
-		}
-
-		return $template;
-	}
-}
-
-
-/**
- * Retrieves a list of unified template objects based on a query.
- *
- * @param array $query {
- *     Optional. Arguments to retrieve templates.
- *
- *     @type array  $slug__in  List of slugs to include.
- *     @type int    $wp_id     Post ID of customized template.
- *     @type string $area      A 'wp_template_part_area' taxonomy value to filter by (for wp_template_part template type only).
- *     @type string $post_type Post type to get the templates for.
- * }
- * @param array $template_type wp_template or wp_template_part.
- *
- * @return array Templates.
- */
-function gutenberg_get_block_templates( $query = array(), $template_type = 'wp_template' ) {
-	/**
-	 * Filters the block templates array before the query takes place.
-	 *
-	 * Return a non-null value to bypass the WordPress queries.
-	 *
-	 * @since 10.8
-	 *
-	 * @param Gutenberg_Block_Template[]|null $block_templates Return an array of block templates to short-circuit the default query,
-	 *                                                  or null to allow WP to run it's normal queries.
-	 * @param array $query {
-	 *     Optional. Arguments to retrieve templates.
-	 *
-	 *     @type array  $slug__in List of slugs to include.
-	 *     @type int    $wp_id Post ID of customized template.
-	 *     @type string $post_type Post type to get the templates for.
-	 * }
-	 * @param array $template_type wp_template or wp_template_part.
-	 */
-	$templates = apply_filters( 'pre_get_block_templates', null, $query, $template_type );
-	if ( ! is_null( $templates ) ) {
-		return $templates;
-	}
-
-	$post_type     = isset( $query['post_type'] ) ? $query['post_type'] : '';
-	$wp_query_args = array(
-		'post_status'    => array( 'auto-draft', 'draft', 'publish' ),
-		'post_type'      => $template_type,
-		'posts_per_page' => -1,
-		'no_found_rows'  => true,
-		'tax_query'      => array(
-			array(
-				'taxonomy' => 'wp_theme',
-				'field'    => 'name',
-				'terms'    => wp_get_theme()->get_stylesheet(),
-			),
-		),
-	);
-
-	if ( 'wp_template_part' === $template_type && isset( $query['area'] ) ) {
-		$wp_query_args['tax_query'][]           = array(
-			'taxonomy' => 'wp_template_part_area',
-			'field'    => 'name',
-			'terms'    => $query['area'],
-		);
-		$wp_query_args['tax_query']['relation'] = 'AND';
-	}
-
-	if ( isset( $query['slug__in'] ) ) {
-		$wp_query_args['post_name__in'] = $query['slug__in'];
-	}
-
-	// This is only needed for the regular templates/template parts CPT listing and editor.
-	if ( isset( $query['wp_id'] ) ) {
-		$wp_query_args['p'] = $query['wp_id'];
-	} else {
-		$wp_query_args['post_status'] = 'publish';
-	}
-
-	$template_query = new WP_Query( $wp_query_args );
-	$query_result   = array();
-	foreach ( $template_query->posts as $post ) {
-		$template = _build_block_template_result_from_post( $post );
-
-		if ( is_wp_error( $template ) ) {
-			continue;
-		}
-
-		if ( $post_type && ! $template->is_custom ) {
-			continue;
-		}
-
-		$query_result[] = $template;
-	}
-
-	if ( ! isset( $query['wp_id'] ) ) {
-		$template_files = _get_block_templates_files( $template_type );
-		foreach ( $template_files as $template_file ) {
-			$template = _build_block_template_result_from_file( $template_file, $template_type );
-
-			if ( $post_type && ! $template->is_custom ) {
-				continue;
-			}
-
-			if ( $post_type &&
-				isset( $template->post_types ) &&
-				! in_array( $post_type, $template->post_types, true )
-			) {
-				continue;
-			}
-
-			$is_not_custom   = false === array_search(
-				wp_get_theme()->get_stylesheet() . '//' . $template_file['slug'],
-				array_column( $query_result, 'id' ),
-				true
-			);
-			$fits_slug_query =
-				! isset( $query['slug__in'] ) || in_array( $template_file['slug'], $query['slug__in'], true );
-			$fits_area_query =
-				! isset( $query['area'] ) || $template_file['area'] === $query['area'];
-			$should_include  = $is_not_custom && $fits_slug_query && $fits_area_query;
-			if ( $should_include ) {
-				$query_result[] = $template;
-			}
-		}
-	}
-	/**
-	 * Filters the array of queried block templates array after they've been fetched.
-	 *
-	 * @since 10.8
-	 *
-	 * @param Gutenberg_Block_Template[] $query_result Array of found block templates.
-	 * @param array $query {
-	 *     Optional. Arguments to retrieve templates.
-	 *
-	 *     @type array  $slug__in List of slugs to include.
-	 *     @type int    $wp_id Post ID of customized template.
-	 * }
-	 * @param array $template_type wp_template or wp_template_part.
-	 */
-	return apply_filters( 'get_block_templates', $query_result, $query, $template_type );
-}
-
-/**
- * Retrieves a single unified template object using its id.
- *
- * @param string $id Template unique identifier (example: theme_slug//template_slug).
- * @param array  $template_type wp_template or wp_template_part.
- *
- * @return Gutenberg_Block_Template|null Template.
- */
-function gutenberg_get_block_template( $id, $template_type = 'wp_template' ) {
-	/**
-	 * Filters the block template object before the query takes place.
-	 *
-	 * Return a non-null value to bypass the WordPress queries.
-	 *
-	 * @since 10.8
-	 *
-	 * @param Gutenberg_Block_Template|null $block_template Return block template object to short-circuit the default query,
-	 *                                               or null to allow WP to run it's normal queries.
-	 * @param string $id Template unique identifier (example: theme_slug//template_slug).
-	 * @param array  $template_type wp_template or wp_template_part.
-	 */
-	$block_template = apply_filters( 'pre_get_block_template', null, $id, $template_type );
-	if ( ! is_null( $block_template ) ) {
-		return $block_template;
-	}
-
-	$parts = explode( '//', $id, 2 );
-	if ( count( $parts ) < 2 ) {
-		return null;
-	}
-	list( $theme, $slug ) = $parts;
-	$wp_query_args        = array(
-		'post_name__in'  => array( $slug ),
-		'post_type'      => $template_type,
-		'post_status'    => array( 'auto-draft', 'draft', 'publish', 'trash' ),
-		'posts_per_page' => 1,
-		'no_found_rows'  => true,
-		'tax_query'      => array(
-			array(
-				'taxonomy' => 'wp_theme',
-				'field'    => 'name',
-				'terms'    => $theme,
-			),
-		),
-	);
-	$template_query       = new WP_Query( $wp_query_args );
-	$posts                = $template_query->posts;
-
-	if ( count( $posts ) > 0 ) {
-		$template = _build_block_template_result_from_post( $posts[0] );
-
-		if ( ! is_wp_error( $template ) ) {
-			return $template;
-		}
-	}
-
-	$block_template = get_block_file_template( $id, $template_type );
-
-	/**
-	 * Filters the queried block template object after it's been fetched.
-	 *
-	 * @since 10.8
-	 *
-	 * @param Gutenberg_Block_Template|null $block_template The found block template, or null if there isn't one.
-	 * @param string $id Template unique identifier (example: theme_slug//template_slug).
-	 * @param array  $template_type wp_template or wp_template_part.
-	 */
-	return apply_filters( 'get_block_template', $block_template, $id, $template_type );
-}
-
 if ( ! function_exists( 'get_block_file_template' ) ) {
 	/**
 	 * Retrieves a single unified template object using its id.
